Lock navmesh for writing only when carving changes, not every frame
Using a NativeDetour to detect when carving NavMeshObstacles are committing changes to the navmesh, we can wrap such changes in calls to NavMeshLock.BeginWrite() and NavMeshLock.EndWrite(), preventing any job threads from reading it during the modification.

The function we are wrapping is NavMeshCarving::ApplyCarvingResults(), which iterates over a field containing a vector of CarveData, in which each element has been previously sent off to a job thread to calculate the modifications that must be made to the navmesh on the main thread. The vector appears as two pointers: one pointing to the start of its allocated buffer, followed by one pointing to the end of the vector after all valid elements.