Use better elevator traversal cost calculations
This should help allow elevators to more accurately determine how long it takes to call and ride them, so that the chosen path is actually the fastest one.

When riding the elevator, the path will not be allowed to change the elevator's destination to avoid the elevator queuing a change in destination after the AI finishes its ride.