Place the destination of FindStraightPath onto the last path polygon
Without moving the end position provided to FindStraightPath onto the provided path, the output corners will end with the destination regardless of whether the path was complete or partial. This would cause pathfinding jobs to think a path was complete when it was not.

The float3 parameters for the method are also changed to Vector3, as it appears the engine itself assumes the memory layout of Vector3 matches the struct in the C++ source.