Make the smart link names per-job instead of a static list
This list would become incorrect as soon as the links in the level changed, like when changing moons. To allow the debug logs to remain accurate after this, the link names are copied into an array and sent off to the individual jobs for them to reference.

This is now only done in debug builds, as the link names are not needed in release.