Add a patch to coalesce all navmesh locks during AI updates
We hook NavMeshManager::Update() to check whether NavMeshSurface, NavMeshObstacle, NavMeshLink, or OffMeshLink need to make modifications to the navmesh, and if so, lock the navmesh for the duration of the update. This ensures that we don't lock and unlock multiple times in a frame, reducing the time wasted in waiting for jobs to pause.