Throw an exception if a smart path task is not started
Since all SmartPathTask methods dealing with path indices would in theory always throw an exception if a task was started with 0 paths, the behavior when starting a task with 0 paths should result in an exception. This also ensures that we don't get into a situation where StartPathTask() is called without actually starting a task, causing IsResultReady() to always return false, which could result in callers being stuck polling it indefinitely.